% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalDEscore.R
\name{CalDEscore}
\alias{CalDEscore}
\title{CalDEscore}
\usage{
CalDEscore(exp, Label)
}
\arguments{
\item{exp}{A gene expression profile of interest (rows are genes, columns are samples).}

\item{Label}{A character vector consist of "0" and "1" which represent sample class in the gene expression profile. "0" means normal sample and "1" means disease sample.}
}
\value{
A matrix with one column of zscore.
}
\description{
Function "CalDEscore" uses gene expression to calculate differential expression level.
}
\examples{
# Obtain the example data
GEP<-Gettest("GEP")
label<-Gettest("label")
# Run the function
DEscore<-CalDEscore(GEP,label)
}
