% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\docType{package}
\name{datasummary}
\alias{DrugSim2DR}
\alias{DrugSim2DR-package}
\title{datasummary: Custom Data Summaries}
\description{
Easily generate custom data frame summaries
}
\author{
\strong{Maintainer}: Junwei Han \email{hanjunwei1981@163.com} [copyright holder]

Authors:
\itemize{
  \item Ji Li
  \item Jiashuo Wu
}

}
