% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gettest.R
\name{Gettest}
\alias{Gettest}
\title{Gettest}
\usage{
Gettest(exampleData)
}
\arguments{
\item{exampleData}{A character,should be one of"Jaccard","commongenes","GO_MF","Drugs","Drugbankid_CID","drugname","GEP","label"}
}
\value{
data
}
\description{
Get the example data
}
