% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myenv}
\alias{myenv}
\title{An environment variable which includes some example data}
\format{
An environment variable
}
\usage{
myenv
}
\description{
An environment variable which includes some example data.
Jaccard:A matrix of Jaccard score between drugs and GOMF.
commongenes:A matrix consisting of genes shared by drug targets and GOMF.
GO_MF:GO terms of molecular functions.
Drugs:Drugs and corresponding targets.
GEP:An example gene expression profile.
label:A vector representing the label of the sample of GEP, where "1" is the disease sample and "0" is the normal sample.
Drugbankid_CID:A dataframe including three columns which are drugbankid, ChembleID, and drugname.
}
\keyword{datasets}
