% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDrugstructure.R
\name{plotDrugstructure}
\alias{plotDrugstructure}
\title{plotDrugstructure}
\usage{
plotDrugstructure(drugid = "")
}
\arguments{
\item{drugid}{A drugbank ID.}
}
\value{
A chemical structure of specific drug
}
\description{
The function "plotDrugstructure" can plot the chemical structure of a drug.
}
\examples{
# Load depend package
library(ChemmineR)
library(rvest)
# Obtain molecular formula and visualize it.
\donttest{plotDrugstructure(drugid ="DB00780")}
}
